+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
-- Use Right and Left arrows without control to scroll query result columns.
+- Use Right and Left arrows without Ctrl to scroll query result columns.
 
 ## [0.1.13] - 2023-11-11